Rewrite to run deploy script on each host
Needed to change the logic to run the deploy script on each host in
sequence. In addition a pre-deploy script is run on the master node
before running the deploy script on each slave. This allows e.g. Django
migrations to be run before deploying new code.

Also set VIRTUAL_ENV when running scripts.